First pass on supporting Apache log format
First, compat with the existing JSON logging is maintained.
This JSON logging feature may be deprecated in the future is favor of a more
powerful one. But this is not the purpose of this changeset.

Then, the goal is to reused the well known Apache log formatting syntax to
define Go-Json-Rest access log.

The user can define his own record format, or pick a predefined one.

This is an implementation of a subset of the Apache mod_log_config syntax,
a lot if still missing. The support can grow over time, but this is already
better than the current solution that offer no customization at all.

See http://httpd.apache.org/docs/2.0/mod/mod_log_config.html for reference.