Introduce shortcuts to create Routes
The four new methods Get, Post, Put and Delete are simple constructors
for the Route objects.

Intantiating the Route object directly is still supported, and allows
the framework to handle any HTTP method.

These shortcuts are there to save keystrokes, improve readability and
to play nicely with golint. (which likes to have key typed in the literals)