Include charset=utf-8 in the JSON response.
Per spec, UTF-8 is the default, and the charset parameter should not
be necessary. But some clients (eg: Chrome) think otherwise.
Since json.Marshal produces UTF-8, setting the charset parameter is a
safe option.

This changeset also includes a better ContentTypeIsJson test method.
It expects the charset to be utf-8 or not set.